Started Work On CommunicationActivity
Added a beginning Layout for CommunicationActivity as well as several fields
to identify elements which may be changed depending on who the user wants
to communicate with. Phone number autolink needs to be figured out and a way
to get a particular employee's contact info needs to be coded, as are
the implicit intents to launch other communication app which are not phone
or email.